usb_kbd: Remove the initial interrupt request
Some keyboards do not send any kind of event after initialization
until the first key press.  The linux kernel initializes keyboards
and lets the interrupt call backs deal with events as they arrive.

Many newer model wireless keyboards act this way.

Without the patch u-boot prints the following on one of keyboards and
leaves it unable to accept input events.

=====
scanning bus xhci_pci for devices...
  Failed to get keyboard state from device 14dd:1009
=====

With the patch, all families of the Raspberry Pi boards can use this
keyboard device.  The patch removes the initial request to the look
for an event immediately.  Because the code that was removed was in
the end of a multi #ifdef #elif, the original printf and error return
needed to be moved directly into the #ifdef
CONFIG_SYS_USB_EVENT_POLL_VIA_INT_QUEUE, to preserve the original
functions and improve readability of the code.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>