xhci-ring.c: Add poll pending state to properly abort transactions
Both xhci_ctrl_tx() and xhci_bulk_tx() can be called synchronously by
other drivers such as the usb storage or network, while the keyboard
driver exclusively uses the polling mode.  The polling mode can start
a transaction, which may then need to be aborted if some other
synchronous driver uses the controller.

Any pending polling transactions must be aborted before switching
modes to avoid corrupting the state of the controller and the driver
which expects a series of commands and responses from a specific
device.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>