xhci-ring.c: Fix crash when issuing "usb reset"
When TRB_TRANSFER is set for a call to xhci_wait_for_event, it can
return null, which causes u-boot to crash.  This is an intermittent
problem found during "usb reset" testing.

If the null return occurs it means there was a timeout, and the
abort_td() should return immediately vs crashing u-boot from a null
dereference.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>