dwc2: Retry when a usb_stall occurs.
Through testing and debugging with various keyboards it was found that
after a usb_port_reset is issued, devices that do not respond soon
enough cause the stall bit to set in the interrupt register.  In the
existing code this will return and error and the device will be
disabled.

Occasionally the Raspberry Pi3 would hang while scanning the USB
devices after a power on reset and this change appears to have
resolved the problem.

The original manifestation of the problem is shown below:

   starting USB...
   Bus usb@7e980000: USB DWC2
   scanning bus usb@7e980000 for devices...
         USB device not accepting new address (error=0)

After the message prints the device is no longer usable.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>